+++ lib/llm/tool/param.rb
@@ -11,7 +11,8 @@ class LLM::Tool
   #   class Greeter < LLM::Tool
   #     name "greeter"
   #     description "Greets the user"
-  #     param :name, String, "The user's name", required: true
+  #     parameter :name, String, "The user's name"
+  #     required %i[name]
   #
   #     def call(name:)
   #       puts "Hello, #{name}!"
@@ -41,6 +42,19 @@ class LLM::Tool
         end
       end
     end
+    alias_method :parameter, :param
+
+    ##
+    # Mark one or more existing parameters as required.
+    # @param names [Array<Symbol,String>]
+    # @return [LLM::Schema::Object]
+    def required(*names)
+      lock do
+        function.params.tap do |schema|
+          names.flatten.each { Utils.fetch(schema.properties, _1).required }
+        end
+      end
+    end
 
     ##
     # @api private
@@ -67,6 +81,10 @@ class LLM::Tool
         leaf.default(default) if default
         leaf.enum(*enum) if enum
         leaf
+      end
+
+      def fetch(properties, name)
+        properties[name] || properties.fetch(name.to_s)
       end
     end
   end